# BLS optimisations