--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -19,5 +19,6 @@ For information on changes in released versions of Teku, see the [releases page]
 ### Additions and Improvements
 - Check `Eth1Address` checksum ([EIP-55](https://eips.ethereum.org/EIPS/eip-55)) if address is mixed-case.
 - Ignore aggregate attestation and sync contribution gossip that does not include any new validators.
+- Optimised BLS batch validation
 
 ### Bug Fixes