## PR Description
 * Use IntStream instead of Stream<Integer>
 * Check semi-aggregate validity and cast in a single pass to avoid multiple iterations of the list


```
Before:
Benchmark                                                    (sigCnt)   Mode  Cnt    Score   Error  Units
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing         1  thrpt   10  814.197 ± 9.578  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing         2  thrpt   10  545.097 ± 9.821  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing         4  thrpt   10  329.395 ± 4.027  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing         8  thrpt   10  185.532 ± 1.362  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing        16  thrpt   10   97.682 ± 1.210  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing        32  thrpt   10   49.997 ± 0.597  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing        64  thrpt   10   25.756 ± 0.360  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing       128  thrpt   10   13.026 ± 0.113  ops/s


After:
Benchmark                                                    (sigCnt)   Mode  Cnt    Score   Error  Units
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing         1  thrpt   10  834.915 ± 7.373  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing         2  thrpt   10  562.865 ± 2.036  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing         4  thrpt   10  337.563 ± 4.653  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing         8  thrpt   10  189.030 ± 1.254  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing        16  thrpt   10  100.144 ± 0.421  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing        32  thrpt   10   51.771 ± 0.263  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing        64  thrpt   10   25.992 ± 1.811  ops/s
BLSBenchmark.verifySignatureBatchedNonParallelDoublePairing       128  thrpt   10   13.231 ± 0.337  ops/s
```


## Documentation

- [x] I thought about documentation and added the `doc-change-required` label to this PR if updates are required.

## Changelog

- [x] I thought about adding a changelog entry, and added one if I deemed necessary.